修复bug以及测试

+++ HonorKingServer/NetFramework/ClientState.cs
@@ -19,7 +19,9 @@ public class ClientState
     /// 客户端的缓冲区
     /// </summary>
     public ByteArray readBuffer = new ByteArray();
-
-    public float lastPingTime = 0;
+    /// <summary>
+    /// 上一次收到ping的时间
+    /// </summary>
+    public long lastPingTime = 0;
 }
